services/financial-analysis-service/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
redis==5.0.1
psycopg2-binary==2.9.9

# Data processing
pandas==2.1.4
numpy==1.25.2
yfinance==0.2.28
alpha-vantage==2.3.1
fredapi==0.5.1

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2
pytest-cov==4.1.0

# Development dependencies
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Monitoring and logging
structlog==23.2.0
prometheus-client==0.19.0

# Security
cryptography>=42.0.0
